Rename mocks to polyfills
